Development 1
For the first development day, I created an initial foundation for the game. In it, I've started a function which shows e-waste moving across the screen and the player losing health once it reaches the end of it's path. Initially I had a barrier of allowing user input while the enemy was moving across the screen in the for loop, however, I've created the event.get() code as a function which I was then able to put inside the for loop within the enemiesMoving function, allowing user input while the enemy is moving.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>