Development 2
Change log: Added recyclers that can be placed along the pathway which damage the waste that goes through it. There is now a health bar of the e-waste shown. If the e-waste loses health, it is deleted. If it reaches the end, it is also deleted and the player loses health.

The code has also been streamlined by removing problematic for loop in enemiesMoving (this was used for x position of the e-waste) that prevented more than one enemy being shown on screen at a time. We now use lists for the x position of the recycler and e-waste, along with their attributes such as health and damage.

The code for drawing map was moved outside of enemiesMoving as it is not related to the animation for the enemies moving.

Also added "import os" and another line of code on line 75 so the game launches in a fixed position.