Development 5: Recyclers use currency, and can not be layered
Changelog:
Green coins are now displayed in the top right.
At the start of the round, players are given a 10 coin bonus.
Recyclers now use a button to be placed. Once placed, the user is charged 10 coins.
Recyclers can no longer be stacked on top of each other.
Start scoreboard menu.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>