Development 11: Scoreboard Design, and a HUD for the game.
Changelog:
Added graphics to scoreboardMenu.
Limit the scores to being only the top 10, and format the numbers shown to be three digits.
Added HUD to gameMenu, which now shows score.
Added a red bar behind the health bar, which shows the health the user lost as the user's health bar shrinks.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>