Development 12: Automatically start next round, and visual improvements.
Changelog:
Move startButton to the rest of the buttons.
Allow the user to automatically start the next round by clicking on it again, or change his mind by clicking once more.
Display roundNumber and status to the HUD. Current status shows when a round has started (10 coin bonus), if a user is trying to place a recycler in a spot where there already is one, or when a user buys/sells a recycler.
Add visual image to each tiered e-Waste, which includes a new list named eWastePic.
Streamline the code for selling recycler by simplifying redundant code with the function, sellingRecycler.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>