Development 14: Tier 2 recyclers, and more!
Changelog:
Add a design to the tutorialMenu, alongside a back button.
Allow the user to exit the program.
Add a secondary tier 1 e-waste that has a 50% chance of spawning.
Streamline code by removing redundant mx = 0 and my = 0.
Add tier 2 recycler that repels the enemy backwards upon contact. This tier has a healthbar and can be damaged.
Add text to most buttons for quality of life.
Prevent the user from being able to place recycler's on the edges of the screen.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>