Development 15: The final update!
Changelog:
Added leave text to exit button and a text that displays the user's placement in scoreboardMenu.
Added a button which allows the user to change his mind on placing or selling a recycler.
Added theme music to main menu.
Added resetVariables function which makes the game fresh when it's being played again, resetting the previous user's data.

Signed-off-by: GamingDevilsCC <gamingdevilscc@gmail.com>